backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
requests==2.31.0
google-generativeai==0.3.2
openai==1.6.1
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.4
